policy: Change Identity and Nexthdr types
Key Identity and Nexthdr fields are now defined as
identity.NumericIdentity and u8proto.U8proto, respectively. This improves
type safety and almost eliminates the need for type casts regarding
these.

+++ pkg/endpoint/endpoint.go
@@ -787,7 +787,7 @@ func (e *Endpoint) Allows(id identity.NumericIdentity) bool {
 	e.unconditionalRLock()
 	defer e.runlock()
 
-	keyToLookup := policy.IngressL3OnlyKey(uint32(id))
+	keyToLookup := policy.IngressL3OnlyKey(id)
 
 	v, ok := e.desiredPolicy.GetPolicyMap().Get(keyToLookup)
 	return ok && !v.IsDeny
